requirements.txt
pytorch-lightning==1.7.7
allennlp==2.9.3
pandas==1.5.0
colorlog==6.7.0
torchmetrics==0.10.0
overrides==7.3.1